Add Vercel logo, NextJS page, layout, and NextJS logo
Added Vercel logo SVG, page.tsx, layout.tsx, NextJS logo SVG and package-lock.json and package.json files. This was necessary to set up the basic page which contains a detailed explanation of Next.js features. The Vercel logo SVG and NextJS logo SVG are used as part of the branding for the app and package.json manages the dependencies of our app.