Update openai's usage in codes and conversation routes
Change the usage of openai in 'CodePage' and 'conversationPage' in dashboard routes. It includes replacing 'ChatCompletionMessage' with 'ChatCompletionRequestMessage' and the replacement of 'openai' package calls. The 'openai' package version is also updated in package.json. The commit also modifies the 'Configuration' and 'createChatCompletion' instances in 'code' and 'conversation' routes in the api. This change was required because the existing calls and instances were outdated and causing issues in the functioning of the app.